Split addProduct and fetchProduct components
