+++ ui/src/App.tsx
@@ -23,20 +23,6 @@ function App() {
       {userID ? (
         <div>
           <ProductPage />
-          <div className="flex">
-            <a href="https://vite.dev" target="_blank">
-              <img src={viteLogo} className="logo" alt="Vite logo" />
-            </a>
-            <a href="https://react.dev" target="_blank">
-              <img src={reactLogo} className="logo react" alt="React logo" />
-            </a>
-          </div>
-          <h1>Vite + React</h1>
-          <div className="card">
-            <button onClick={() => setCount((count) => count + 1)}>
-              count is {count}
-            </button>
-          </div>
         </div>
       ) : (
         <UserInfo userID={userID} setUserID={setUserID} />
